Implemented base functionality. Clean up and improvements to come next.
Removed dependency on depercated gets. Deprecated in c++14.
Implemented moderen c++ method for reading STDIN cin. Cin by default works with c++ types. Removed excessive uses of c-types to c++ types to maintain redability and reduce risk of type casting mistakes and buffer errors.
All types are strings now. Can be moved back to char arrays if needed for space but recommend sticking to vectors verses c-style null terminated. This will reduce chances of potential buffer issues.

Its not stated the nature of the words in the requirements so sticking to arbitrary length. If sticking to words within a language the longest single word in English is 45 characters, 64 in Icelandic, or 180k for a protein name and that can be used as a buffer. Unless we decide to include compound words.
Because of the uncertainty with these requirements strings seem the safest approach. This is where buisness case would influence design decisions ie. a word list for a writing program, a word list for a protein catalog.

With the above changes the need for the Word struct went away so it was removed.