Cleaned up code and have the most basic version of the requirements completed.
Will accept only one input per line. Note does not account for words with spaces and drops inputs after the first word's space.

Uses a map to maintin alphabetic order and quick addition and removal of words. Uses a queue to add words to be processed into map(a little overkill but so is the thread itself). Mutex's added to ensure thread safety. Also added delete to clean up thread.

Next move is to change it to break up files, switch out to smart pointers, and add testing.